requirements.txt
anyio==3.6.1
bcrypt==3.2.2
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.2
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
fastapi==0.78.0
h11==0.13.0
idna==3.3
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tomli==2.0.1
typing-extensions==4.3.0
urllib3==1.26.9
uvicorn==0.18.2
